refactor(structure): 도메인 및 예외 처리 클래스의 패키지 경로 변경
- global 도메인과 member 도메인 관련 클래스들이 새로운 패키지 구조로 이동되었습니다.
- BaseEntity, DomainException, ImageContentType 클래스등이 각 도메인 별로 적절한 위치로 이동하였습니다.
- member 도메인 관련 EncryptedPassword, Nickname, Password VO 둥 클래스들이 새로운 위치로 변경되었습니다.
- 이동된 클래스들의 패키지 참조가 업데이트 되었으며, 관련 예외 처리 클래스들도 새로운 패키지 구조에 맞게 조정되었습니다.
- 이 구조 변경을 통해 프로젝트의 모듈화 및 도메인 별 관리가 용이해질 것으로 예상됩니다.
