+++ src/main/java/sidepair/persistence/member/MemberRepository.java
@@ -2,9 +2,9 @@ package sidepair.persistence.member;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
-import sidepair.member.domain.Member;
-import sidepair.member.domain.vo.Email;
-import sidepair.member.domain.vo.Nickname;
+import sidepair.domain.member.Member;
+import sidepair.domain.member.vo.Email;
+import sidepair.domain.member.vo.Nickname;
 
 public interface MemberRepository extends JpaRepository<Member, Long>, MemberQueryRepository {
