+++ src/main/java/sidepair/service/dto/auth/LoginDto.java
@@ -1,7 +1,7 @@
 package sidepair.service.dto.auth;
 
-import sidepair.member.domain.vo.Email;
-import sidepair.member.domain.vo.Password;
+import sidepair.domain.member.vo.Email;
+import sidepair.domain.member.vo.Password;
 
 public record LoginDto(
         Email email,
