+++ src/main/java/sidepair/service/dto/mamber/OauthMemberJoinDto.java
@@ -1,4 +1,4 @@
-package sidepair.member.configuration.dto;
+package sidepair.service.dto.mamber;
 
 import sidepair.service.dto.mamber.request.PositionType;
