+++ src/main/java/sidepair/service/member/MemberService.java
@@ -1,4 +1,4 @@
-package sidepair.member.application;
+package sidepair.service.member;
 
 
 import java.net.URL;
@@ -9,16 +9,16 @@ import org.springframework.core.env.Environment;
 import org.springframework.http.HttpMethod;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import sidepair.global.domain.ImageContentType;
-import sidepair.global.domain.NumberGenerator;
-import sidepair.member.configuration.dto.MemberSkillDto;
-import sidepair.member.configuration.dto.MemberSkillSaveDto;
-import sidepair.member.configuration.dto.OauthMemberJoinDto;
-import sidepair.member.domain.MemberSkill;
-import sidepair.member.domain.MemberSkills;
-import sidepair.member.domain.Position;
-import sidepair.member.domain.vo.Nickname;
-import sidepair.member.domain.vo.SkillName;
+import sidepair.domain.ImageContentType;
+import sidepair.service.NumberGenerator;
+import sidepair.service.dto.mamber.MemberSkillDto;
+import sidepair.service.dto.mamber.MemberSkillSaveDto;
+import sidepair.service.dto.mamber.OauthMemberJoinDto;
+import sidepair.domain.member.MemberSkill;
+import sidepair.domain.member.MemberSkills;
+import sidepair.domain.member.Position;
+import sidepair.domain.member.vo.Nickname;
+import sidepair.domain.member.vo.SkillName;
 import sidepair.persistence.auth.RefreshTokenRepository;
 import sidepair.service.FileService;
 import sidepair.service.aop.ExceptionConvert;
@@ -26,19 +26,19 @@ import sidepair.service.auth.TokenProvider;
 import sidepair.service.dto.auth.response.AuthenticationResponse;
 import sidepair.service.exception.ConflictException;
 import sidepair.service.exception.NotFoundException;
-import sidepair.member.application.mapper.MemberMapper;
-import sidepair.member.configuration.dto.MemberInformationDto;
-import sidepair.member.configuration.dto.MemberInformationForPublicDto;
-import sidepair.member.configuration.request.MemberJoinRequest;
-import sidepair.member.configuration.dto.MemberJoinDto;
-import sidepair.member.configuration.response.MemberInformationForPublicResponse;
-import sidepair.member.configuration.response.MemberInformationResponse;
-import sidepair.member.domain.EncryptedPassword;
-import sidepair.member.domain.Member;
-import sidepair.member.domain.MemberProfile;
-import sidepair.member.domain.vo.MemberImage;
+import sidepair.service.mapper.MemberMapper;
+import sidepair.service.dto.mamber.MemberInformationDto;
+import sidepair.service.dto.mamber.MemberInformationForPublicDto;
+import sidepair.service.dto.mamber.request.MemberJoinRequest;
+import sidepair.service.dto.mamber.MemberJoinDto;
+import sidepair.service.dto.mamber.response.MemberInformationForPublicResponse;
+import sidepair.service.dto.mamber.response.MemberInformationResponse;
+import sidepair.domain.member.EncryptedPassword;
+import sidepair.domain.member.Member;
+import sidepair.domain.member.MemberProfile;
+import sidepair.domain.member.vo.MemberImage;
 import sidepair.persistence.member.MemberRepository;
-import sidepair.member.domain.vo.Email;
+import sidepair.domain.member.vo.Email;
 import sidepair.service.mapper.AuthMapper;
 
 @Service
