+++ src/test/java/sidepair/domain/member/vo/PasswordTest.java
@@ -1,11 +1,11 @@
-package sidepair.member.domain.vo;
+package sidepair.domain.member.vo;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import sidepair.member.exception.MemberException;
+import sidepair.domain.member.exception.MemberException;
 
 class PasswordTest {
