+++ src/test/java/sidepair/persistence/member/MemberRepositoryTest.java
@@ -5,21 +5,20 @@ import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
-import sidepair.global.domain.ImageContentType;
-import sidepair.member.domain.EncryptedPassword;
-import sidepair.member.domain.Member;
-import sidepair.member.domain.MemberProfile;
-import sidepair.member.domain.MemberSkill;
-import sidepair.member.domain.MemberSkills;
-import sidepair.member.domain.Position;
-import sidepair.member.domain.vo.Email;
-import sidepair.member.domain.vo.MemberImage;
-import sidepair.member.domain.vo.Nickname;
-import sidepair.member.domain.vo.Password;
-import sidepair.member.domain.vo.SkillName;
+import sidepair.domain.ImageContentType;
+import sidepair.domain.member.EncryptedPassword;
+import sidepair.domain.member.Member;
+import sidepair.domain.member.MemberProfile;
+import sidepair.domain.member.MemberSkill;
+import sidepair.domain.member.MemberSkills;
+import sidepair.domain.member.Position;
+import sidepair.domain.member.vo.Email;
+import sidepair.domain.member.vo.MemberImage;
+import sidepair.domain.member.vo.Nickname;
+import sidepair.domain.member.vo.Password;
+import sidepair.domain.member.vo.SkillName;
 import sidepair.persistence.helper.RepositoryTest;
 
 @RepositoryTest
